Simple request routing
Addresses the basic needs of request routing (issue #14), but the
routing is still very minimal and not super friendly from a coding
perspective. This will be remedied in the future.

+++ SwiftCGI Framework/SwiftCGI/Record.swift
@@ -59,7 +59,7 @@ class FCGIRecord {
     var fcgiPacketData: NSData {
         var bytes = [UInt8](count: 8, repeatedValue: 0)
         bytes[0] = version.rawValue
-        bytes[1] = type.rawValue
+        bytes[1] = UInt8(type.rawValue)
         
         let (msb, lsb) = requestID.decomposeBigEndian()
         bytes[2] = msb
